Fix CPAN Bug #76445
t/06-cfgver.t fails because of using wrong perl executable
----
The t/06-cfgver.t test fails on most systems, see
http://matrix.cpantesters.org/ ?dist=Config-Versioned-0.5 . That's
because bin/cfgver script is used here, which is probably using the
system perl, and not the perl used for building and testing
Config-Versioned. It's best to use something like

    $^X bin/cfgver

instead, so you make sure that the correct perl executable is in use.